pyproject.toml
[project]
name = "cdk_tools"
version = "0.1.0"
description = "Shared base classes and utilities for AWS CDK projects."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "aws-cdk-lib>=2.0.0,<3.0.0",
    "constructs>=10.0.0,<11.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
